+++ src/yt/types.rs
@@ -1,8 +1,8 @@
 #![allow(dead_code)]
 
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct VideoDetails {
     pub id: String,
     pub title: String,